Rename ride paint header files
