+++ src/openrct2/ride/coaster/BolligerMabillardTrack.h
