+++ src/openrct2/ride/coaster/JuniorRollerCoaster.h
